backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
